rm poisson vestiges; dedupe dgp; delete unused effects methods
