+++ bayespecon/__init__.pyi
@@ -28,9 +28,6 @@ from .dgp.flows import (
 from .dgp.flows import (
     generate_panel_negbin_flow_data_separable as generate_panel_negbin_flow_data_separable,
 )
-from .dgp.flows import (
-    generate_poisson_flow_data as generate_poisson_flow_data,
-)
 from .diagnostics import (
     ModelComparison as ModelComparison,
 )
@@ -115,15 +112,6 @@ from .models import (
 )
 from .models import (
     OLSPanelRE as OLSPanelRE,
-)
-from .models import (
-    PoissonFlowPanel as PoissonFlowPanel,
-)
-from .models import (
-    PoissonSARFlowPanel as PoissonSARFlowPanel,
-)
-from .models import (
-    PoissonSARFlowSeparablePanel as PoissonSARFlowSeparablePanel,
 )
 from .models import (
     SARFlowPanel as SARFlowPanel,
@@ -226,15 +214,6 @@ from .models.flow import (
 )
 from .models.flow import (
     OLSFlow as OLSFlow,
-)
-from .models.flow import (
-    PoissonFlow as PoissonFlow,
-)
-from .models.flow import (
-    PoissonSARFlow as PoissonSARFlow,
-)
-from .models.flow import (
-    PoissonSARFlowSeparable as PoissonSARFlowSeparable,
 )
 from .models.flow import (
     SARFlow as SARFlow,
