+++ bayespecon/dgp/__init__.pyi
@@ -44,22 +44,10 @@ from .flows import (
     generate_panel_negbin_flow_data_separable as generate_panel_negbin_flow_data_separable,
 )
 from .flows import (
-    generate_panel_poisson_flow_data as generate_panel_poisson_flow_data,
-)
-from .flows import (
-    generate_panel_poisson_flow_data_separable as generate_panel_poisson_flow_data_separable,
-)
-from .flows import (
     generate_panel_sem_flow_data as generate_panel_sem_flow_data,
 )
 from .flows import (
     generate_panel_sem_flow_data_separable as generate_panel_sem_flow_data_separable,
-)
-from .flows import (
-    generate_poisson_flow_data as generate_poisson_flow_data,
-)
-from .flows import (
-    generate_poisson_flow_data_separable as generate_poisson_flow_data_separable,
 )
 from .flows import (
     generate_sem_flow_data as generate_sem_flow_data,
