+++ bayespecon/dgp/panel_tobit.py
@@ -6,19 +6,8 @@ with fixed-effects transformations.
 
 from __future__ import annotations
 
-import numpy as np
-
 from .panel_fe import simulate_panel_sar_fe, simulate_panel_sem_fe
-from .utils import make_panel_output_geodataframe
-
-
-def _left_censor(
-    y_latent: np.ndarray, censoring: float
-) -> tuple[np.ndarray, np.ndarray]:
-    mask = y_latent <= censoring
-    y_obs = y_latent.copy()
-    y_obs[mask] = censoring
-    return y_obs, mask
+from .utils import _left_censor, make_panel_output_geodataframe
 
 
 def simulate_panel_sar_tobit_fe(
