+++ bayespecon/models/cross_section/ols.py
@@ -89,21 +89,6 @@ class OLS(GaussianLikelihoodMixin, SpatialModel):
     _model_type: str = "ols"
     _likelihood: str = "gaussian"  # NUTS-only (no _gibbs_key)
 
-    def _compute_spatial_effects(self) -> dict[str, np.ndarray]:
-        """Not applicable — OLS has no spatial lag structure.
-
-        Raises
-        ------
-        NotImplementedError
-            Always raised; use Bayesian LM diagnostics instead to
-            assess spatial structure after estimation.
-        """
-        raise NotImplementedError(
-            "OLS has no spatial lag structure and therefore no spatial effects. "
-            "Use Bayesian LM diagnostics to assess which spatial model "
-            "is appropriate, then refit with SAR, SEM, SLX, SDM, or SDEM."
-        )
-
     def _compute_spatial_effects_posterior(
         self,
     ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
